move client interactions to separate page
