+++ frontend/public/locales/ua.json
@@ -404,7 +404,8 @@
     "patients": {
       "create": "Зареєструвати нового клієнта",
       "details": "Деталі клієнта",
-      "conversations": "Підсумок бесід"
+      "conversations": "Підсумок бесід",
+      "interactions": "Взаємодії з клієнтом"
     },
     "chatbot": {
       "create": "Створити нового чат-бота",
